requirements.txt
anaconda-clean==1.0
anaconda-client==1.5.1
anaconda-navigator==1.3.1
argcomplete==1.0.0
astropy==1.2.1
Babel==2.3.4
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.5.1
conda==4.2.9
conda-build==2.0.2
Flask==0.11.1
Flask-Cors==2.1.2
future==0.16.0
geopy==1.11.0
gitdb2==2.0.0
GitPython==2.1.3
googlefinance==0.7
ipykernel==4.5.0
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
Jinja2==2.8
nb-anacondacloud==1.2.0
nb-conda==2.0.0
nb-conda-kernels==2.0.0
numpy==1.11.1
oauth2==1.9.0.post1
oauthlib==2.0.2
openpyxl==2.3.2
py==1.4.31
pyasn1==0.1.9
pycosat==0.6.1
pycparser==2.14
pycrypto==2.6.1
pycurl==7.43.0
pyflakes==1.3.0
Pygments==2.1.3
pylint==1.5.4
pyOpenSSL==16.0.0
pyparsing==2.1.4
pytest==2.9.2
python-dateutil==2.5.3
python-dotenv==0.6.4
python-forecastio==1.3.5
python-twitter==3.2.1
pytz==2016.6.1
pywin32==220
PyYAML==3.12
pyzmq==15.4.0
QtAwesome==0.3.3
qtconsole==4.2.1
QtPy==1.1.2
requests==2.11.1
requests-oauthlib==0.8.0
responses==0.5.1
rope-py3k==0.9.4.post1
ruamel-yaml===-VERSION
scikit-image==0.12.3
scikit-learn==0.17.1
scipy==0.18.1
sympy==1.0
tables==3.2.2
toolz==0.8.0
